More platform information code moved from shared to platform.
