+++ libraries/shared/src/SharedUtil.cpp
@@ -27,7 +27,6 @@
 
 #ifdef Q_OS_WIN
 #include <windows.h>
-#include "CPUIdent.h"
 #include <Psapi.h>
 
 #if _MSC_VER >= 1900
@@ -190,7 +189,7 @@ void outputBits(unsigned char byte, QDebug* continuedDebug) {
         resultString.sprintf("[ %d (0x%x): ", byte, byte);
     }
     debug << qPrintable(resultString);
-    
+
     for (int i = 0; i < 8; i++) {
         resultString.sprintf("%d", byte >> (7 - i) & 1);
         debug << qPrintable(resultString);
@@ -738,7 +737,7 @@ QString formatSecTime(qint64 secs) {
 QString formatSecondsElapsed(float seconds) {
     QString result;
 
-    const float SECONDS_IN_DAY = 60.0f * 60.0f * 24.0f;        
+    const float SECONDS_IN_DAY = 60.0f * 60.0f * 24.0f;
     if (seconds > SECONDS_IN_DAY) {
         float days = floor(seconds / SECONDS_IN_DAY);
         float rest = seconds - (days * SECONDS_IN_DAY);
@@ -781,252 +780,6 @@ void disableQtBearerPoll() {
     const QByteArray DISABLE_BEARER_POLL_TIMEOUT = QString::number(-1).toLocal8Bit();
     qputenv("QT_BEARER_POLL_TIMEOUT", DISABLE_BEARER_POLL_TIMEOUT);
 }
-
-void printSystemInformation() {
-    // Write system information to log
-    qCDebug(shared) << "Build Information";
-    qCDebug(shared).noquote() << "\tBuild ABI: " << QSysInfo::buildAbi();
-    qCDebug(shared).noquote() << "\tBuild CPU Architecture: " << QSysInfo::buildCpuArchitecture();
-
-    qCDebug(shared).noquote() << "System Information";
-    qCDebug(shared).noquote() << "\tProduct Name: " << QSysInfo::prettyProductName();
-    qCDebug(shared).noquote() << "\tCPU Architecture: " << QSysInfo::currentCpuArchitecture();
-    qCDebug(shared).noquote() << "\tKernel Type: " << QSysInfo::kernelType();
-    qCDebug(shared).noquote() << "\tKernel Version: " << QSysInfo::kernelVersion();
-
-    qCDebug(shared) << "\tOS Version: " << QOperatingSystemVersion::current();
-
-#ifdef Q_OS_WIN
-    SYSTEM_INFO si;
-    GetNativeSystemInfo(&si);
-
-    qCDebug(shared) << "SYSTEM_INFO";
-    qCDebug(shared).noquote() << "\tOEM ID: " << si.dwOemId;
-    qCDebug(shared).noquote() << "\tProcessor Architecture: " << si.wProcessorArchitecture;
-    qCDebug(shared).noquote() << "\tProcessor Type: " << si.dwProcessorType;
-    qCDebug(shared).noquote() << "\tProcessor Level: " << si.wProcessorLevel;
-    qCDebug(shared).noquote() << "\tProcessor Revision: "
-                       << QString("0x%1").arg(si.wProcessorRevision, 4, 16, QChar('0'));
-    qCDebug(shared).noquote() << "\tNumber of Processors: " << si.dwNumberOfProcessors;
-    qCDebug(shared).noquote() << "\tPage size: " << si.dwPageSize << " Bytes";
-    qCDebug(shared).noquote() << "\tMin Application Address: "
-                       << QString("0x%1").arg(qulonglong(si.lpMinimumApplicationAddress), 16, 16, QChar('0'));
-    qCDebug(shared).noquote() << "\tMax Application Address: "
-                       << QString("0x%1").arg(qulonglong(si.lpMaximumApplicationAddress), 16, 16, QChar('0'));
-
-    const double BYTES_TO_MEGABYTE = 1.0 / (1024 * 1024);
-
-    qCDebug(shared) << "MEMORYSTATUSEX";
-    MEMORYSTATUSEX ms;
-    ms.dwLength = sizeof(ms);
-    if (GlobalMemoryStatusEx(&ms)) {
-        qCDebug(shared).noquote()
-            << QString("\tCurrent System Memory Usage: %1%").arg(ms.dwMemoryLoad);
-        qCDebug(shared).noquote()
-            << QString("\tAvail Physical Memory: %1 MB").arg(ms.ullAvailPhys * BYTES_TO_MEGABYTE, 20, 'f', 2);
-        qCDebug(shared).noquote()
-            << QString("\tTotal Physical Memory: %1 MB").arg(ms.ullTotalPhys * BYTES_TO_MEGABYTE, 20, 'f', 2);
-        qCDebug(shared).noquote()
-            << QString("\tAvail in Page File:    %1 MB").arg(ms.ullAvailPageFile * BYTES_TO_MEGABYTE, 20, 'f', 2);
-        qCDebug(shared).noquote()
-            << QString("\tTotal in Page File:    %1 MB").arg(ms.ullTotalPageFile * BYTES_TO_MEGABYTE, 20, 'f', 2);
-        qCDebug(shared).noquote()
-            << QString("\tAvail Virtual Memory:  %1 MB").arg(ms.ullAvailVirtual * BYTES_TO_MEGABYTE, 20, 'f', 2);
-        qCDebug(shared).noquote()
-            << QString("\tTotal Virtual Memory:  %1 MB").arg(ms.ullTotalVirtual * BYTES_TO_MEGABYTE, 20, 'f', 2);
-    } else {
-        qCDebug(shared) << "\tFailed to retrieve memory status: " << GetLastError();
-    }
-
-    qCDebug(shared) << "CPUID";
-
-    qCDebug(shared) << "\tCPU Vendor: " << CPUIdent::Vendor().c_str();
-    qCDebug(shared) << "\tCPU Brand:  " << CPUIdent::Brand().c_str();
-
-    for (auto& feature : CPUIdent::getAllFeatures()) {
-        qCDebug(shared).nospace().noquote() << "\t[" << (feature.supported ? "x" : " ") << "] " << feature.name.c_str();
-    }
-#endif
-
-    qCDebug(shared) << "Environment Variables";
-    // List of env variables to include in the log. For privacy reasons we don't send all env variables.
-    const QStringList envWhitelist = {
-        "QTWEBENGINE_REMOTE_DEBUGGING"
-    };
-    auto envVariables = QProcessEnvironment::systemEnvironment();
-    for (auto& env : envWhitelist)
-    {
-        qCDebug(shared).noquote().nospace() << "\t" <<
-            (envVariables.contains(env) ? " = " + envVariables.value(env) : " NOT FOUND");
-    }
-}
-
-bool getMemoryInfo(MemoryInfo& info) {
-#ifdef Q_OS_WIN
-    MEMORYSTATUSEX ms;
-    ms.dwLength = sizeof(ms);
-    if (!GlobalMemoryStatusEx(&ms)) {
-        return false;
-    }
-
-    info.totalMemoryBytes = ms.ullTotalPhys;
-    info.availMemoryBytes = ms.ullAvailPhys;
-    info.usedMemoryBytes = ms.ullTotalPhys - ms.ullAvailPhys;
-
-
-    PROCESS_MEMORY_COUNTERS_EX pmc;
-    if (!GetProcessMemoryInfo(GetCurrentProcess(), reinterpret_cast<PROCESS_MEMORY_COUNTERS*>(&pmc), sizeof(pmc))) {
-        return false;
-    }
-    info.processUsedMemoryBytes = pmc.PrivateUsage;
-    info.processPeakUsedMemoryBytes = pmc.PeakPagefileUsage;
-
-    return true;
-#endif
-
-    return false;
-}
-
-// Largely taken from: https://msdn.microsoft.com/en-us/library/windows/desktop/ms683194(v=vs.85).aspx
-
-#ifdef Q_OS_WIN
-using LPFN_GLPI = BOOL(WINAPI*)(
-    PSYSTEM_LOGICAL_PROCESSOR_INFORMATION,
-    PDWORD);
-
-DWORD CountSetBits(ULONG_PTR bitMask)
-{
-    DWORD LSHIFT = sizeof(ULONG_PTR) * 8 - 1;
-    DWORD bitSetCount = 0;
-    ULONG_PTR bitTest = (ULONG_PTR)1 << LSHIFT;
-    DWORD i;
-
-    for (i = 0; i <= LSHIFT; ++i) {
-        bitSetCount += ((bitMask & bitTest) ? 1 : 0);
-        bitTest /= 2;
-    }
-
-    return bitSetCount;
-}
-#endif
-
-bool getProcessorInfo(ProcessorInfo& info) {
-
-#ifdef Q_OS_WIN
-    LPFN_GLPI glpi;
-    bool done = false;
-    PSYSTEM_LOGICAL_PROCESSOR_INFORMATION buffer = NULL;
-    PSYSTEM_LOGICAL_PROCESSOR_INFORMATION ptr = NULL;
-    DWORD returnLength = 0;
-    DWORD logicalProcessorCount = 0;
-    DWORD numaNodeCount = 0;
-    DWORD processorCoreCount = 0;
-    DWORD processorL1CacheCount = 0;
-    DWORD processorL2CacheCount = 0;
-    DWORD processorL3CacheCount = 0;
-    DWORD processorPackageCount = 0;
-    DWORD byteOffset = 0;
-    PCACHE_DESCRIPTOR Cache;
-
-    glpi = (LPFN_GLPI)GetProcAddress(
-        GetModuleHandle(TEXT("kernel32")),
-        "GetLogicalProcessorInformation");
-    if (nullptr == glpi) {
-        qCDebug(shared) << "GetLogicalProcessorInformation is not supported.";
-        return false;
-    }
-
-    while (!done) {
-        DWORD rc = glpi(buffer, &returnLength);
-
-        if (FALSE == rc) {
-            if (GetLastError() == ERROR_INSUFFICIENT_BUFFER) {
-                if (buffer) {
-                    free(buffer);
-                }
-
-                buffer = (PSYSTEM_LOGICAL_PROCESSOR_INFORMATION)malloc(
-                    returnLength);
-
-                if (NULL == buffer) {
-                    qCDebug(shared) << "Error: Allocation failure";
-                    return false;
-                }
-            } else {
-                qCDebug(shared) << "Error " << GetLastError();
-                return false;
-            }
-        } else {
-            done = true;
-        }
-    }
-
-    ptr = buffer;
-
-    while (byteOffset + sizeof(SYSTEM_LOGICAL_PROCESSOR_INFORMATION) <= returnLength) {
-        switch (ptr->Relationship) {
-        case RelationNumaNode:
-            // Non-NUMA systems report a single record of this type.
-            numaNodeCount++;
-            break;
-
-        case RelationProcessorCore:
-            processorCoreCount++;
-
-            // A hyperthreaded core supplies more than one logical processor.
-            logicalProcessorCount += CountSetBits(ptr->ProcessorMask);
-            break;
-
-        case RelationCache:
-            // Cache data is in ptr->Cache, one CACHE_DESCRIPTOR structure for each cache. 
-            Cache = &ptr->Cache;
-            if (Cache->Level == 1) {
-                processorL1CacheCount++;
-            } else if (Cache->Level == 2) {
-                processorL2CacheCount++;
-            } else if (Cache->Level == 3) {
-                processorL3CacheCount++;
-            }
-            break;
-
-        case RelationProcessorPackage:
-            // Logical processors share a physical package.
-            processorPackageCount++;
-            break;
-
-        default:
-            qCDebug(shared) << "\nError: Unsupported LOGICAL_PROCESSOR_RELATIONSHIP value.\n";
-            break;
-        }
-        byteOffset += sizeof(SYSTEM_LOGICAL_PROCESSOR_INFORMATION);
-        ptr++;
-    }
-
-    qCDebug(shared) << "GetLogicalProcessorInformation results:";
-    qCDebug(shared) << "Number of NUMA nodes:" << numaNodeCount;
-    qCDebug(shared) << "Number of physical processor packages:" << processorPackageCount;
-    qCDebug(shared) << "Number of processor cores:" << processorCoreCount;
-    qCDebug(shared) << "Number of logical processors:" << logicalProcessorCount;
-    qCDebug(shared) << "Number of processor L1/L2/L3 caches:"
-        << processorL1CacheCount
-        << "/" << processorL2CacheCount
-        << "/" << processorL3CacheCount;
-
-    info.numPhysicalProcessorPackages = processorPackageCount;
-    info.numProcessorCores = processorCoreCount;
-    info.numLogicalProcessors = logicalProcessorCount;
-    info.numProcessorCachesL1 = processorL1CacheCount;
-    info.numProcessorCachesL2 = processorL2CacheCount;
-    info.numProcessorCachesL3 = processorL3CacheCount;
-
-    free(buffer);
-
-    return true;
-#endif
-
-    return false;
-}
-
 
 const QString& getInterfaceSharedMemoryName() {
     static const QString applicationName = "High Fidelity Interface - " + qgetenv("USERNAME");
